+++ src/MicroElements.Functional/Messages/MessageSeverity.cs
@@ -1,0 +1,23 @@
+﻿namespace MicroElements.Functional
+{
+    /// <summary>
+    /// Message severity.
+    /// </summary>
+    public enum MessageSeverity
+    {
+        /// <summary>
+        /// Information message.
+        /// </summary>
+        Information,
+
+        /// <summary>
+        /// Warning.
+        /// </summary>
+        Warning,
+
+        /// <summary>
+        /// Error message.
+        /// </summary>
+        Error,
+    }
+}